Move the semaphore lock in anthropic's api client to outside of the retry logic
As written, this code does not properly implement backoff retry logic when more requests to the client are made than there are num_threads. You can see this in the following example:

* Queue 100 API calls with 10 concurrents (i.e., self.available_threads=10)
* This semaphore locks, so only 10 calls are made
* All 10 of these calls return 529 **and immediately exit the semaphore** before being put to sleep by asyncio
* 10 more calls are made immediately, ignoring the asyncio sleep time
* This repeats, usually leading to no backoff between retries

The new code (which moves the semaphore lock up) fixes this by only allowing num_threads requests to be in the calling and waiting states combined.

I've tested this manually on my local machine, but do not have any unit tests for this.